fixing package issues, refactoring
